# cmd/contentmanagement/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorBody:
    properties:
      fieldName:
        type: string
      message:
        description: 'required: true'
        type: string
    type: object
  models.GenericError:
    properties:
      body:
        $ref: '#/definitions/models.ErrorBody'
        description: 'in: body'
      statusCode:
        description: swagger:ignore
        type: integer
    type: object
...
  workspace.CreateWorkspaceRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
  workspace.Tag:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  workspace.Workspace:
    properties:
      id:
        type: string
      languages:
        items:
          type: string
        type: array
      name:
        type: string
      tags:
        items:
          $ref: '#/definitions/workspace.Tag'
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /contentmanagement/workspace:
    post:
      description: Create a new workspace
      parameters:
      - description: workspace body
        in: body
        name: workspace
        required: true
        schema:
          $ref: '#/definitions/workspace.CreateWorkspaceRequest'
      responses:
        "201":
          description: Created
          headers:
            Location:
              type: string
          schema:
            $ref: '#/definitions/workspace.Workspace'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.GenericError'
      summary: Create workspace
      tags:
      - workspace
  /contentmanagement/workspace/{id}:
    get:
      description: Get workspace by id
      parameters:
      - description: uuid formatted ID.
        format: uuid
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/workspace.Workspace'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.GenericError'
      summary: Get workspace
      tags:
      - workspace
    put:
      description: Update a new workspace
      parameters:
      - description: uuid formatted ID.
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: workspace body
        in: body
        name: workspace
        required: true
        schema:
          $ref: '#/definitions/workspace.Workspace'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/workspace.Workspace'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.GenericError'
      summary: Update workspace
      tags:
      - workspace
...
swagger: "2.0"